+++ src/main.rs
@@ -156,7 +156,7 @@ fn add_and_commit(repo: &Repository, path: &Path, message: &str) -> Result<Oid, 
     index.write()?;
     let oid = index.write_tree()?;
     let tree = repo.find_tree(oid)?;
-    let config = git2::Config::open_default()?;
+    let config = repo.config()?;
     let signature = git2::Signature::now(
         config.get_entry("user.name")?.value().unwrap(),
         config.get_entry("user.email")?.value().unwrap(),